Enum flac::ErrorKind

pub enum ErrorKind {
    IO(ErrorKind),
    Incomplete(usize),
    Continue,
    EndOfInput,
    Unknown,
    HeaderParser,
    MetadataHeaderParser,
    StreamInfoParser,
    PaddingParser,
    ApplicationParser,
    SeekTableParser,
    VorbisCommentParser,
    CueSheetParser,
    PictureParser,
    UnknownParser,
    BlockingStrategyParser,
    BlockingSampleParser,
    ChannelBitsParser,
    UTF8HeaderParser,
    UTF8BodyParser,
    BlockSizeParser,
    SampleRateParser,
    CRC8Parser,
    FrameFooterParser,
    SubframeHeaderParser,
    LeadingZerosParser,
    ConstantParser,
    VerbatimParser,
    FixedParser,
    LPCParser,
    InvalidBlockType,
    InvalidSyncCode,
    InvalidBlockSample,
    InvalidChannelBits,
    InvalidUTF8,
    InvalidCRC8,
    InvalidCRC16,
    InvalidSubframeHeader,
    NotFound,
}

Represent the different kinds of errors.

Variants

IO

Error from I/O.

Incomplete

A parser stopped midway and need more bytes to consume.

Continue

A parser has completes and there is still more bytes to consume.

EndOfInput

A parser has completes and there is no more bytes to consume.

Unknown

A non-specified error.

HeaderParser

Failed parsing the "fLaC" header token.

MetadataHeaderParser

Failed parsing a metadata header.

StreamInfoParser

Failed parsing the metadata block StreamInfo.

PaddingParser

Failed parsing the metadata block Padding.

ApplicationParser

Failed parsing the metadata block Application.

SeekTableParser

Failed parsing the metadata block SeekTable.

VorbisCommentParser

Failed parsing the metadata block VorbisComment.

CueSheetParser

Failed parsing the metadata block CueSheet.

PictureParser

Failed parsing the metadata block Picture.

UnknownParser

Failed parsing the metadata block Unknown.

BlockingStrategyParser

Failed parsing the blocking strategy inside the frame header.

BlockingSampleParser

Failed parsing the blocking sample inside the frame header.

ChannelBitsParser

Failed parsing the channel bits inside the frame header.

UTF8HeaderParser

Failed parsing the UTF-8 header inside the frame header.

UTF8BodyParser

Failed parsing the UTF-8 body inside the frame header.

BlockSizeParser

Failed parsing the secondary block size inside the frame header.

SampleRateParser

Failed parsing the secondary sample rate inside the frame header.

CRC8Parser

Failed parsing the CRC-8 inside the frame header.

FrameFooterParser

Failed parsing the frame footer, also known as the CRC-16.

SubframeHeaderParser

Failed parsing the subframe header.

LeadingZerosParser

Failed parsing the leading zero for a unary value.

ConstantParser

Failed parsing a Constant subframe data.

VerbatimParser

Failed parsing a Verbatim subframe data.

FixedParser

Failed parsing a Fixed subframe data.

LPCParser

Failed parsing a LPC subframe data.

InvalidBlockType

A block type, base on the number, that is outside the range (0-126).

InvalidSyncCode

An incorrect sync code with the frame header.

InvalidBlockSample

A block sample that could cause sync-fooling.

InvalidChannelBits

One or more bits are reserved values.

InvalidUTF8

An error occurred in building the UTF-8 value.

InvalidCRC8

The stored CRC-8 doesn't match the one generated from the bytes within the frame header.

InvalidCRC16

The stored CRC-16 doesn't match the one generated from the bytes within the entire frame.

InvalidSubframeHeader

A subframe header that could cause sync-fooling.

NotFound

Some metadata block was not found with a specific filter.