Module flac::frame::subframe

Structs

EntropyCodingMethod
Header for the entropy coding method.

Fixed
Fixed linear prediction subframe.

LPC
Finite Impulse Response (FIR) linear prediction subframe.

PartitionedRice
Header for a Rice partitioned residual.

PartitionedRiceContents
Contents of a Rice partitioned residual.

SubFrame
A single channel of audio data.

Enums

CodingMethod
The available entropy coding methods.

Data
General enum that holds all the different subframe data types.

Constants

MAX_FIXED_ORDER
Maximum order of the fixed predictors permitted by the format.

MAX_LPC_ORDER
Maximum LPC order permitted by the format.

Functions

subframe_parser
Parse a single channel of audio data.