Module flac::metadata

Provides an interface for dealing with FLAC metadata blocks.

Structs

Application
Data used by third-party applications.

CueSheet
Stores cue information.

CueSheetTrack
Track information inside a cue sheet.

CueSheetTrackIndex
An index point within a track, inside of a cue sheet.

Metadata
Data associated with a single metadata block.

Picture
Stores pictures associated with the FLAC file.

SeekPoint
Seek, or skip, to a point within the FLAC file.

StreamInfo
Information regarding the entire audio stream.

VorbisComment
Stores human-readable name/value pairs.

Enums

Data
General enum that hold all the different metadata block data.

PictureType
The picture type according to the ID3v2 attached picture frame.

Type
An enum that represents a metadata block type.

Functions

get_cue_sheet
Reads and returns the CueSheet metadata block of the given FLAC file.

get_picture
Reads and returns a Picture metadata block of the given FLAC file.

get_stream_info
Reads and returns the StreamInfo metadata block of the given FLAC file.

get_vorbis_comment
Reads and returns the VorbisComment metadata block of the given FLAC file.

metadata_parser
Parse a metadata block.