Enum flac::metadata::Data

pub enum Data {
    StreamInfo(StreamInfo),
    Padding(u32),
    Application(Application),
    SeekTable(Vec<SeekPoint>),
    VorbisComment(VorbisComment),
    CueSheet(CueSheet),
    Picture(Picture),
    Unknown(Vec<u8>),
}

General enum that hold all the different metadata block data.

Variants

StreamInfo

Information regarding the entire audio stream.

Padding

Block that represents a number of padded bytes.

Application

Data used by third-party applications.

SeekTable

Table of multiple points to seek, or skip, to within the FLAC file.

VorbisComment

Stores human-readable name/value pairs.

CueSheet

Stores cue information

Picture

Stores pictures associated with the FLAC file.

Unknown

A type of block data that isn't know or doesn't match the type above.