Struct flac::Metadata

pub struct Metadata {
    pub data: Data,
    // some fields omitted
}

Data associated with a single metadata block.

Fields

data

Block data containing one of the eight different types of metadata.

Methods

impl Metadata

fn new(is_last: bool, length: u32, data: Data) -> Self

Constructs a new Metadata struct based on the arguments passed in.

fn is_last(&self) -> bool

Returns whether the current metadata block is the last.

fn data_type(&self) -> Type

Returns the metadata block's type.

fn is_stream_info(&self) -> bool

Returns true when the current Metadata is StreamInfo.

fn is_padding(&self) -> bool

Returns true when the current Metadata is Padding.

fn is_application(&self) -> bool

Returns true when the current Metadata is Application.

fn is_seek_table(&self) -> bool

Returns true when the current Metadata is SeekTable.

fn is_vorbis_comment(&self) -> bool

Returns true when the current Metadata is VorbisComment.

fn is_cue_sheet(&self) -> bool

Returns true when the current Metadata is CueSheet.

fn is_picture(&self) -> bool

Returns true when the current Metadata is Picture.

fn is_unknown(&self) -> bool

Returns true when the current Metadata is Unknown.

fn bytes_len(&self) -> usize

fn to_bytes<Write: Write>(&self, buffer: &mut Write) -> Result<()>